Fix deactivate failure, backend coverage
